pyproject.toml
[project]
name = "ytscript"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "build>=1.2.2.post1",
    "huggingface-hub>=0.29.3",
    "nuitka>=2.6.8",
    "openai-whisper>=20240930",
    "pytest>=8.3.5",
    "torch>=2.6.0",
    "transformers>=4.50.2",
    "yt-dlp>=2025.3.27",
    "accelerate>=0.26.0",
]
